+++ lib/index.jsx
@@ -1,20 +1,26 @@
-// Bundle tailwind
-import "./css/tailwind.css";
-
-// Bundle groundwork-water styles
-import "@usace/groundwork/dist/style.css";
-import "react-toastify/dist/ReactToastify.css";
-
 // Import components
 import TSTable from "./components/data/tables/TSTable";
 import CWMSTable from "./components/data/tables/CWMSTable";
+import {
+  buildCsvContent,
+  buildSeriesLookup,
+  buildTableIndex,
+  buildTableRows,
+  buildTableRowValues,
+  downloadBlob,
+} from "./components/data/tables/tableData";
 import GageMap from "./components/data/maps/GageMap";
 import CWMSPlot from "./components/data/plots/CWMSPlot";
+import BasinPie from "./components/data/plots/BasinPie";
+import RadialFillChart from "./components/data/plots/RadialFillChart";
 import CdaLatestValueCard from "./components/data/cards/CdaLatestValueCard";
+import DamProfile from "./components/data/plots/dam-profile/chart";
+import DataStatus from "./components/data/summary/DataStatus";
 
 // Import input components
 import CWMSInput from "./components/data/forms/inputs/CWMSInput";
 import CWMSFileUpload from "./components/data/forms/inputs/CWMSFileUpload";
+import CWMSDataUpload from "./components/data/forms/inputs/CWMSDataUpload";
 import CWMSTextarea from "./components/data/forms/inputs/CWMSTextarea";
 import CWMSCheckboxes from "./components/data/forms/inputs/CWMSCheckboxes";
 import CWMSRadioGroup from "./components/data/forms/inputs/CWMSRadioGroup";
@@ -51,6 +57,12 @@ import useCdaTimeSeriesGroup from "./components/data/hooks/useCdaTimeSeriesGroup
 import useCdaOffices from "./components/data/hooks/useCdaOffices";
 import useNwpsGauge from "./components/data/hooks/useNwpsGauge";
 import useNwpsGaugeData from "./components/data/hooks/useNwpsGaugeData";
+import useDataStatusFile from "./components/data/hooks/useDataStatusFile";
+import useCwmsDataUpload from "./components/data/hooks/useCwmsDataUpload";
+import {
+  fetchCdaLevelTimeSeries,
+  fetchCdaLevelValues,
+} from "./components/data/helpers/levels";
 
 // files
 import useCdaBlob from "./components/data/hooks/useCdaBlob";
@@ -62,6 +74,16 @@ import {
   getLocationCatalogCoordinates,
   locationCatalogToFeatureCollection,
 } from "./components/data/maps/locationCatalog";
+import {
+  CWMS_DATA_UPLOAD_HEADERS,
+  CwmsDataUploadValidationError,
+  buildCwmsDataUploadPayloads,
+  classifyCwmsDataUploadRows,
+  createCwmsDataUploadTemplate,
+  filterCwmsDataUploadRows,
+  parseCwmsDataUploadRows,
+  readCwmsDataUploadFile,
+} from "./components/data/forms/helpers/dataUpload";
 // Utility Functions
 import {
   PRECISION_BY_UNIT,
@@ -83,10 +105,19 @@ import { OfficeDropdown } from "./components/data/dropdowns/OfficeDropdown";
 export {
   TSTable,
   CWMSTable,
+  buildCsvContent,
+  buildSeriesLookup,
+  buildTableIndex,
+  buildTableRows,
+  buildTableRowValues,
+  downloadBlob,
   GageMap,
   CWMSPlot,
+  BasinPie,
+  RadialFillChart,
   CdaLatestValueCard,
   CdaUrlProvider,
+  DataStatus,
   OfficeDropdown,
   useCdaBlob,
   useCdaBlobs,
@@ -99,9 +130,21 @@ export {
   useCdaMultiTimeSeries,
   useCdaTimeSeriesGroup,
   useCdaOffices,
+  useDataStatusFile,
+  useCwmsDataUpload,
+  fetchCdaLevelTimeSeries,
+  fetchCdaLevelValues,
   useDebounce,
   getLocationCatalogCoordinates,
   locationCatalogToFeatureCollection,
+  CWMS_DATA_UPLOAD_HEADERS,
+  CwmsDataUploadValidationError,
+  buildCwmsDataUploadPayloads,
+  classifyCwmsDataUploadRows,
+  createCwmsDataUploadTemplate,
+  filterCwmsDataUploadRows,
+  parseCwmsDataUploadRows,
+  readCwmsDataUploadFile,
   useNwpsGauge,
   useNwpsGaugeData,
   AuthProvider,
@@ -114,6 +157,7 @@ export {
   // Input components
   CWMSInput,
   CWMSFileUpload,
+  CWMSDataUpload,
   CWMSTextarea,
   CWMSCheckboxes,
   CWMSRadioGroup,
@@ -135,5 +179,6 @@ export {
   updateToast,
   formatSubmissionMessage,
   showDetailedError,
+  DamProfile,
 };
 // export { helperFunction };